Split native methods into own file
